pyproject.toml
[tool.poetry]
name = "skypro-django-hw1"
version = "0.1.0"
description = ""
authors = ["Maxxchop <4mu@bk.ru>"]
readme = "README.md"
packages = [{include = "skypro_django_hw1"}]

[tool.poetry.dependencies]
python = "^3.10"
docker = "^6.0.1"
psycopg2 = "^2.9.5"
pillow = "^9.3.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
